os: Resolve warnings in os folder
libxipelf/xipelf.c:103:27: warning: assignment to 'uint32_t' {aka 'unsigned int'} from 'void *' makes integer from pointer without a cast [-Wint-conversion]
libxipelf/xipelf.c:104:27: warning: assignment to 'uint32_t' {aka 'unsigned int'} from 'void *' makes integer from pointer without a cast [-Wint-conversion]
libxipelf/xipelf.c:105:25: warning: assignment to 'uint32_t' {aka 'unsigned int'} from 'void *' makes integer from pointer without a cast [-Wint-conversion]
libxipelf/xipelf.c:106:25: warning: assignment to 'uint32_t' {aka 'unsigned int'} from 'void *' makes integer from pointer without a cast [-Wint-conversion]
libxipelf/xipelf.c:107:23: warning: assignment to 'uint32_t' {aka 'unsigned int'} from 'void *' makes integer from pointer without a cast [-Wint-conversion]
/root/tizenrt/os/board/common/partitions.c:285:7: warning: unused variable 'fs_devname' [-Wunused-variable]
/root/tizenrt/os/board/common/partitions.c:284:6: warning: unused variable 'ret' [-Wunused-variable]
rtl8730e_boot.c:181:6: warning: assignment to 'struct i2c_dev_s *' from 'int' makes pointer from integer without a cast [-Wint-conversion]
rtl8730e_ndp120.c:179:10: warning: assignment to 'struct audio_lowerhalf_s *' from 'int' makes pointer from integer without a cast [-Wint-conversion]
rtl8730e_ndp120.c:145:6: warning: 'ret' may be used uninitialized in this function [-Wmaybe-uninitialized]
audio/syu645bscripts.h:36:1: warning: useless storage class specifier in empty declaration
binary_manager/binary_manager_verify.c:184:15: warning: 'crc_bufsize' may be used uninitialized in this function [-Wmaybe-uninitialized]
log_dump/log_dump.c:231:52: warning: passing argument 2 of 'compress_block' from incompatible pointer type [-Wincompatible-pointer-types]
pm_timedsuspend.c:127:18: warning: passing argument 1 of 'timer_timeout' makes integer from pointer without a cast [-Wint-conversion]
pm_timedsuspend.c:159:17: warning: passing argument 1 of 'timer_timeout' makes integer from pointer without a cast [-Wint-conversion]

+++ os/binfmt/libxipelf/xipelf.c
@@ -100,11 +100,11 @@ static int xipelf_loadbinary(FAR struct binary_s *binp)
 
 	close(filfd);
 
-	binp->sections[BIN_TEXT] = uspace.text_start;
-	binp->flash_region_start = uspace.text_start - uspace_offset + 4;
-	binp->flash_region_end = uspace.flash_end;
-	binp->ram_region_start = uspace.ram_start;
-	binp->ram_region_end = uspace.ram_end;
+	binp->sections[BIN_TEXT] = (uint32_t)uspace.text_start;
+	binp->flash_region_start = (uint32_t)uspace.text_start - uspace_offset + 4;
+	binp->flash_region_end = (uint32_t)uspace.flash_end;
+	binp->ram_region_start = (uint32_t)uspace.ram_start;
+	binp->ram_region_end = (uint32_t)uspace.ram_end;
 	binp->sizes[BIN_TEXT] = binp->flash_region_end - binp->flash_region_start;
 
 	/* zero out the bss... */
